Move rendering tests into their own namespace, delete example test
