+++ database/factories/UserFactory.php
@@ -17,14 +17,16 @@ class UserFactory extends Factory
      */
     public function definition()
     {
+        $username = $this->faker->userName();
+
         return [
-            'username' => $this->faker->userName(),
+            'username' => $username,
             'email' => $this->faker->unique()->safeEmail,
             'email_verified_at' => now(),
             'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
             'remember_token' => Str::random(10),
             'reg_ip' => "127.0.0.1",
-            'api_token' => 'afakeapitoken',
+            'api_token' => "afakeapitoken$username",
         ];
     }
