--- tests/Feature/ExampleTest.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Tests\Feature;
-
-use Tests\TestCase;
-
-class ExampleTest extends TestCase
-{
-    /**
-     * A basic test example.
-     *
-     * @return void
-     */
-    public function testBasicTest()
-    {
-        $response = $this->get('/');
-
-        $response->assertStatus(200);
-    }
-}
