+++ tests/Feature/Rendering/TextRendererTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Rendering;
 
 use App\Models\Calendar;
 use App\Services\RendererService\TextRenderer;
